pyproject.toml
[project]
name = "biosciences-research"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "jupyter>=1.1.1",
    "langchain-experimental>=0.3.4",
    "langchain>=0.3.19",
    "langchain-cohere==0.4.4",
    "cohere==5.12.0",
    "langchain-openai>=0.3.7",
    "langgraph==0.6.7",
    "qdrant-client>=1.17.0",
    "rank-bm25>=0.2.2",
    "langchain-qdrant>=0.2.0",
    "ragas==0.2.10",
    "pymupdf>=1.26.3",
    "rapidfuzz>=3.14.1",
    "claude-agent-sdk>=0.1.1",
    "huggingface-hub>=0.26.0",
    "datasets>=3.2.0",
    "streamlit>=1.40.0",
    "langgraph-cli[inmem]>=0.4.4",
    "openai>=2.4.0",
    "python-dotenv>=1.1.1",
]
